refactored fixed list enums and pebble filters to seperate them, reverted notify template and created ducplicate, included new template id
