+++ e2e/paths/enterOrdersAndDirections_test.js
@@ -1,6 +1,6 @@
 const config = require('../config.js');
 
-Feature('Enter order and details');
+Feature('Enter order and details').retry(2);
 
 Before((I, caseViewPage) => {
   I.logInAndCreateCase(config.swanseaLocalAuthorityEmailUserOne, config.localAuthorityPassword);
