+++ service/src/main/java/uk/gov/hmcts/reform/pebble/OrderMappingFilter.java
@@ -1,13 +1,13 @@
 package uk.gov.hmcts.reform.pebble;
 
 import com.mitchellbosecke.pebble.extension.Filter;
-import uk.gov.hmcts.reform.fpl.config.utils.FixedListMappingType;
+import uk.gov.hmcts.reform.fpl.config.utils.OrderType;
 
 import java.util.List;
 import java.util.Map;
 
 
-public class FixedListMappingFilter implements Filter {
+public class OrderMappingFilter implements Filter {
 
     @Override
     public List<String> getArgumentNames() {
@@ -16,6 +16,6 @@ public class FixedListMappingFilter implements Filter {
 
     @Override
     public Object apply(Object input, Map<String, Object> args) {
-        return FixedListMappingType.valueOf((String) input).getLabel();
+        return OrderType.valueOf((String) input).getLabel();
     }
 }
