+++ service/src/main/resources/HTMLTemplate.html
@@ -206,14 +206,14 @@
       <dd class="data-list__definition">
         {% if case_data.orders is not empty and case_data.orders.orderType is not empty %}
           {% for type in case_data.orders.orderType %}
-            {{ type | fixedlistmapping }}
+            {{ type | ordermapping }}
           {% endfor %}
           {% if case_data.orders.otherOrder is not empty %}
             <span class="u-block">{{ case_data.orders.otherOrder }}</span>
           {% endif %}
           {% if case_data.orders.emergencyProtectionOrders is not empty %}
             {% for order in case_data.orders.emergencyProtectionOrders %}
-              {{ order | fixedlistmapping }}
+              {{ order | emergencyprotectionordermapping }}
             {% endfor %}
             {% if case_data.orders.emergencyProtectionOrderDetails is not empty %}
               <span class="u-block">{{ case_data.orders.emergencyProtectionOrderDetails }}</span>
@@ -231,7 +231,7 @@
           {% if case_data.orders.orderType is not empty or case_data.orders.directions is not empty %}
             {% if case_data.orders.emergencyProtectionOrderDirection is not empty %}
               {% for epo in case_data.orders.emergencyProtectionOrderDirection %}
-                {{ epo | fixedlistmapping }}
+                {{ epo | emergencyprotectionorderdirectionmapping }}
               {% endfor %}
             {% endif %}
             {% if case_data.orders.emergencyProtectionOrderDirectionDetails is not empty %}
